# Optimize unescaping/parsing performance of BamlResourceContent/BamlStringToken