@dipeshmsft 2025-01-03
@h3xds1nz, this looks good to me, probably we can add some tests for this change as soon as the unit test project for PresentationFramework is setup.