@h3xds1nz 2025-01-04
@dipeshmsft No worries. Might not have been the best description I put out first time around haha.

The changes are in the last commit (just 2x `== null` to `is null` and comment spaces).